# Improve Pendulum.now and Pendulum.utcnow performance