@pganssle 2018-01-05
There are also several places in the tests where you use `utcnow()` when `now(UTC)` would work just fine, but I didn't want to touch those because I was not entirely sure if constructing an instance from a naive datetime was actually part of the test for some reason.

It does seem to me that a lot of the tests use `now` or `utcnow()` when they could use datetime literals instead, I'm not sure why that was done that way.