@sdispater 2018-01-19
Thanks for that !

Regarding the use of `now()` or `utcnow()` in the tests, it's just for conciseness more than anything.